Moved all the damage NodeCoupling algorithms from Utilities->Damage. This involved some fiddling with the python bindings.
