+++ src/Damage/DamageModel.cc
@@ -17,13 +17,13 @@
 #include "Boundary/Boundary.hh"
 #include "Kernel/TableKernel.hh"
 #include "Neighbor/ConnectivityMap.hh"
-#include "Utilities/NodeCoupling.hh"
-#include "Utilities/DamagedNodeCoupling.hh"
-#include "Utilities/DamageGradientNodeCoupling.hh"
-#include "Utilities/ThreePointDamagedNodeCoupling.hh"
 #include "Utilities/GeometricUtilities.hh"
 #include "Utilities/safeInv.hh"
 #include "Utilities/Timer.hh"
+#include "Utilities/NodeCoupling.hh"
+#include "Damage/DamagedNodeCoupling.hh"
+#include "Damage/DamageGradientNodeCoupling.hh"
+#include "Damage/ThreePointDamagedNodeCoupling.hh"
 
 #include "boost/shared_ptr.hpp"
