+++ src/Damage/MaxDamageNodeCoupling.hh
