+++ src/Damage/ThreePointDamagedNodeCouplingInst.cc.py
@@ -2,9 +2,9 @@ text = """
 //------------------------------------------------------------------------------
 // Explicit instantiation.
 //------------------------------------------------------------------------------
-#include "Utilities/DamagedNodeCoupling.cc"
+#include "Damage/ThreePointDamagedNodeCoupling.cc"
 
 namespace Spheral {
-  template class DamagedNodeCoupling< Dim< %(ndim)s > >;
+  template class ThreePointDamagedNodeCoupling< Dim< %(ndim)s > >;
 }
 """
