+++ src/Pybind11Wraps/Damage/DamageMOD.py
@@ -13,6 +13,9 @@ from DamageModel import *
 from TensorDamageModel import *
 from ProbabilisticDamageModel import *
 from JohnsonCookDamage import *
+from DamagedNodeCoupling import *
+from DamageGradientNodeCoupling import *
+from ThreePointDamagedNodeCoupling import *
 
 #-------------------------------------------------------------------------------
 # Includes
@@ -26,6 +29,9 @@ PYB11includes += ['"NodeList/SolidNodeList.hh"',
                   '"Damage/weibullFlawDistributionBenzAsphaug.hh"',
                   '"Damage/weibullFlawDistributionOwen.hh"',
                   '"Damage/computeFragmentField.hh"',
+                  '"Damage/DamagedNodeCoupling.hh"',
+                  '"Damage/ThreePointDamagedNodeCoupling.hh"',
+                  '"Damage/DamageGradientNodeCoupling.hh"',
                   '"FileIO/FileIO.hh"']
 
 #-------------------------------------------------------------------------------
@@ -106,6 +112,10 @@ DamageModel%(ndim)id = PYB11TemplateClass(DamageModel, template_parameters="%(Di
 TensorDamageModel%(ndim)id = PYB11TemplateClass(TensorDamageModel, template_parameters="%(Dimension)s")
 ProbabilisticDamageModel%(ndim)id = PYB11TemplateClass(ProbabilisticDamageModel, template_parameters="%(Dimension)s")
 JohnsonCookDamage%(ndim)id = PYB11TemplateClass(JohnsonCookDamage, template_parameters="%(Dimension)s")
+
+DamagedNodeCoupling%(ndim)id = PYB11TemplateClass(DamagedNodeCoupling, template_parameters="%(Dimension)s")
+ThreePointDamagedNodeCoupling%(ndim)id = PYB11TemplateClass(ThreePointDamagedNodeCoupling, template_parameters="%(Dimension)s")
+DamageGradientNodeCoupling%(ndim)id = PYB11TemplateClass(DamageGradientNodeCoupling, template_parameters="%(Dimension)s")
 
 weibullFlawDistributionBenzAsphaug%(ndim)id = PYB11TemplateFunction(weibullFlawDistributionBenzAsphaug, template_parameters="%(Dimension)s")
 weibullFlawDistributionOwen%(ndim)id = PYB11TemplateFunction(weibullFlawDistributionOwen, template_parameters="%(Dimension)s")
